cambios a los integer para recursos con más capacidad

+++ database/migrations/2018_06_29_210447_create_en_construcciones_table.php
@@ -15,7 +15,7 @@ class CreateEnConstruccionesTable extends Migration
     {
         Schema::create('en_construcciones', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('personal');
+            $table->decimal('personal', 14, 2);
             $table->integer('nivel');
             $table->string('accion');
             $table->timestamps();
